refactor: Date -> String으로 반환해주는 로직 usecase로 빼기
