+++ Trinap/Sources/Domain/UseCases/Photographer/Protocols/ConvertDateToStringUseCase.swift
@@ -1,0 +1,13 @@
+//
+//  ConvertDateToStringUseCase.swift
+//  Trinap
+//
+//  Created by kimchansoo on 2022/12/12.
+//  Copyright © 2022 Trinap. All rights reserved.
+//
+
+import Foundation
+
+protocol ConvertDateToStringUseCase {
+    func convert(startDate: Date, endDate: Date) -> String?
+}
