+++ Trinap/Sources/Presenter/TabBar/Main/PhotographerDetail/PhotographerDetailViewController.swift
@@ -289,37 +289,6 @@ extension PhotographerDetailViewController {
             .regular(string: "날짜 변경하기")
         calendarButton.setAttributedTitle(buttonText, for: .normal)
     }
-    
-    private func formattingCalendarButtonText(startDate: Date, endDate: Date) -> String? {
-        let startSeperated = startDate.toString(type: .yearToSecond).components(separatedBy: " ")
-        let endSeperated = endDate.toString(type: .yearToSecond).components(separatedBy: " ")
-        
-        guard let date = startSeperated[safe: 0] else { return nil }
-        let dateSeperated = date.components(separatedBy: "-")
-        guard
-            let month = dateSeperated[safe: 1],
-            let day = dateSeperated[safe: 2]
-        else { return nil }
-        
-        guard
-            let startTime = startSeperated.last,
-            let endTime = endSeperated.last
-        else { return nil }
-        let startHourToSec = startTime.components(separatedBy: ":")
-        let endHourToSec = endTime.components(separatedBy: ":")
-        guard
-            let startHour = startHourToSec[safe: 0],
-            let startMin = startHourToSec[safe: 1],
-            let endHour = endHourToSec[safe: 0],
-            let endMin = endHourToSec[safe: 1]
-        else { return nil }
-
-        let reservationDate = "\(month)/\(day)"
-        let reservationStart = "\(startHour):\(startMin)"
-        let reservationEnd = "\(endHour):\(endMin)"
-        let dateInfo = "\(reservationDate) \(reservationStart)-\(reservationEnd)\n"
-        return dateInfo
-    }
 }
 
 private extension NSMutableAttributedString {
